DXP-ProductBot/requirements.txt
requests
pymsteams
pandas
python-decouple
azure-storage-blob
fastapi
uvicorn